+++ tests/testapp/urls.py
@@ -1,6 +1,7 @@
+from django.conf.urls import include, url
 from django.contrib import admin
-from django.urls import path
 
 urlpatterns = [
-    path('admin/', admin.site.urls),
+    url(r'^admin/', admin.site.urls),
+    url(r'^celery-tracker/', include("django_celery_tracker.urls")),
 ]