+++ CHANGELOG.md
@@ -6,9 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.1.0]
+
 ### Added
 -Dashboard view included in `django_celery_tracker.urls`
--Field `args` to `CeleryTask` model
+-Field `args` to `CeleryTask` model (don't forget to run `./manage.py migrate`)
 
 ## [1.0.1]
 
@@ -18,5 +20,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 -Disclaimer for database usage
 
-[Unreleased]: https://github.com/chris-allen/django-celery-tracker/compare/v1.0.1...HEAD
+[Unreleased]: https://github.com/chris-allen/django-celery-tracker/compare/v1.1.0...HEAD
+[1.0.1]: https://github.com/chris-allen/django-celery-tracker/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/chris-allen/django-celery-tracker/compare/v1.0.0...v1.0.1